// src/api/routeHalte.ts
// Generated by https://quicktype.io

export interface HalteProps {
    periode_data: number;
    wilayah:      string;
    kecamatan:    string;
    kelurahan:    string;
    nama_halte:   string;
    lokasi:       string;
    koordinat_x:  string;
    koordinat_y:  string;
}

// Generated by https://quicktype.io

export interface PenumpangProps {
    periode_data:     number;
    jenis:            string;
    kode_trayek:      number | string;
    trayek:           string;
    jumlah_penumpang: number;
}